Fin implémentation CRUD alleys + alleys orders

+++ app.py
@@ -1,6 +1,6 @@
 import sqlite3
 
-from flask import Flask, render_template, g, jsonify, request
+from flask import Flask, render_template, g, jsonify, request, redirect, url_for
 from flask_cors import CORS
 
 DATABASE = 'groceries.db'
@@ -20,47 +20,67 @@ def close_connection(exception):
     if db is not None:
         db.close()
 
+# Views
 @app.get("/")
 def home():
     return render_template("index.html")
 
-@app.get("/api/alleys")
-def get_all_alleys():
+@app.get("/alleys")
+def alleys():
+    cur = get_db().cursor()
+    cur.execute('SELECT a.rowid, a.name, ao."order" FROM alleys a LEFT JOIN alleys_orders ao ON ao.alley_id = a.rowid;')
+    data = cur.fetchall()
+    alleys = []
+    for i in data:
+            alleys.append({ "id": i[0], "name": i[1], "order": "" if i[2] is None else i[2] })
+    return render_template("alleys.html", alleys=alleys)
+
+@app.route("/alleys/add", methods=["GET", "POST"])
+def alleys_add():
+    if request.method == "POST":
+        alley = request.form["alley"]
+
+        if alley is None or alley == "":
+            return "", 400
+    
+        cur = get_db().cursor()
+        cur.execute("INSERT INTO alleys (name) VALUES (?)", (alley.upper(),))
+        cur.connection.commit()
+
+        return redirect(url_for('alleys'))
+    return render_template("alleys/add.html")
+
+@app.route("/alleys/order", methods=["GET", "POST"])
+def alleys_order():
+    if request.method == "POST":
+        body = request.get_json()
+        cur = get_db().cursor()
+        cur.execute("DELETE FROM alleys_orders;")
+        cur.connection.commit()
+        for item in body:
+            cur.execute('INSERT INTO alleys_orders (alley_id, "order") VALUES (?, ?)', (item["alley_id"],item["order"],))
+            cur.connection.commit()
+        return "", 201
+
     cur = get_db().cursor()
     cur.execute("SELECT rowid, name FROM alleys")
     data = cur.fetchall()
     alleys = []
     for i in data:
             alleys.append({ "id": i[0], "name": i[1] })
-    return jsonify(alleys)
 
-@app.post("/api/alleys")
-def create_alley():
-    body = request.get_json()
-    name = body["name"]
-    cur = get_db().cursor()
-    cur.execute("INSERT INTO alleys (name) VALUES (?)", (name,))
-    cur.connection.commit()
-    return "",201
+    return render_template("alleys/order.html", alleys=alleys)
 
-@app.delete("/api/alleys/<int:id>")
-def delete_alley(id):
+@app.delete("/alleys/delete/<int:id>")
+def alleys_delete(id):
     cur = get_db().cursor()
     cur.execute("DELETE FROM alleys WHERE rowid = ?", (id,))
     cur.connection.commit()
     return "",204
 
-@app.post("/api/alleys_orders")
-def set_alleys_orders():
-    body = request.get_json()
-    cur = get_db().cursor()
-    cur.execute("DELETE FROM alleys_orders")
-    cur.connection.commit()
-    for item in body:
-        cur.execute('INSERT INTO alleys_orders (alley_id, "order") VALUES (?, ?)', (item["id"], item["order"], ))
-        cur.connection.commit()
-        print(item)
-    return "", 201
+@app.get("/products")
+def products():
+    return render_template("products.html")
 
 if __name__ == '__main__':
     app.run(debug=True)
